pyproject.toml
[project]
name = "experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvclive>=3.48.3",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "pyaml>=25.7.0",
    "scikit-learn>=1.7.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
]
